Changed how I'm splitting the scene dialog
